refactor  : 해시태그 검색 부분과 상태바, SearchFilter 컴포넌트를 분리하여 코드수를 줄인다
